+++ js/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Removed identifyBytesFull and identifyStreamFull: identifyBytes and
   identifyStream now return all the scores as well (accessible with
   `result.prediction.scores_map`).
+- Restrict the input types to `Uint8Array` and `Buffer`.
 
 ## [0.2.13] - 2024-03-26